+++ Gems/OpenXRVk/Code/Source/Builders/OpenXRVkAssetBuildersSystemComponent.cpp
