+++ Gems/OpenXRVk/Code/Source/Builders/OpenXRVkAssetBuildersSystemComponent.h
