+++ Gems/OpenXRVk/Code/openxrvk_shared_builder_files.cmake
@@ -7,5 +7,5 @@
 #
 
 set(FILES
-    Source/Builders/OpenXRBuilderModule.cpp
+    Source/Builders/OpenXRVkBuilderModule.cpp
 )
